refactor: market position code tidy up
