+++ programs/monaco_protocol/src/instructions/market_position/mod.rs
@@ -1,7 +1,13 @@
 pub mod create_market_position;
 pub mod settle_market_position;
+pub mod update_on_order_cancellation;
+pub mod update_on_order_creation;
+pub mod update_on_order_match;
 pub mod void_market_position;
 
 pub use create_market_position::*;
 pub use settle_market_position::*;
+pub use update_on_order_cancellation::*;
+pub use update_on_order_creation::*;
+pub use update_on_order_match::*;
 pub use void_market_position::*;
